stage 1

+++ src/main/java/org/arkaan/simpleatm/datamodel/Card.java
@@ -23,7 +23,7 @@ public class Card {
         this.pin = pin;
     }
 
-    public double getAccountBalance() {
+    public Integer getAccountBalance() {
         return customer.getBalance();
     }
