+++ src/main/java/org/arkaan/simpleatm/datamodel/Customer.java
@@ -26,7 +26,7 @@ public class Customer {
         return name;
     }
 
-    double getBalance() {
+    Integer getBalance() {
         return balance;
     }
